+++ fuzzy/heat_bed_test/heatBedFuzzy/fuzzy.h
@@ -17,22 +17,22 @@
 struct membr_set_val
 {
 	short int set_1;
-	short int deg_truth_1;
+	float deg_truth_1;
 
 	short int set_2;
-	short int deg_truth_2;
+	float deg_truth_2;
 };
 
 struct op_membr_val
 {
 	short int set_1;
-	short int deg_truth_1;
+	float deg_truth_1;
 
 	short int set_2;
-	short int deg_truth_2;
+	float deg_truth_2;
 
 	short int set_3;
-	short int deg_truth_3;
+	float deg_truth_3;
 };
 
 class fuzzy
@@ -73,18 +73,18 @@ class fuzzy
 		
 		fuzzy(short int n, int imax, int imin, int omax, int omin);
 
-		short int percentizer	 (int val, int val_max, int val_min);
-		int depercentizer  (int val, int val_max, int val_min);
+		float percentizer	 (int val, int val_max, int val_min);
+		float depercentizer  (int val, int val_max, int val_min);
 		
 		/* calculates error percentized and saves value in error_p*/
-		short int error_calc(int val, int set_val);
-		short int ch_error_calc (int en, int *en_1);	//calculate percentized change of error
+		float error_calc(int val, int set_val);
+		float ch_error_calc (int en, int *en_1);	//calculate percentized change of error
 		
 		struct membr_set_val membership_determiner(short int n, short int val);	//used to determine membership values of a variable
 		
 		struct op_membr_val ch_op_determiner(short int n, struct membr_set_val ip_mem_1, struct membr_set_val ip_mem_2);
 		
-		short int defuzzifier (short int n, struct op_membr_val u);
+		float defuzzifier (short int n, struct op_membr_val u);
 
 		int fuzzy_controller(int input, int s_point);
 };
